+++ frontend/src/app/not-found.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import Link from 'next/link'
-import { Button } from '~/components/ui/buttons/Button'
+import { Button } from '~/components/ui/button'
 
 export default function NotFound() {
   return (
