+++ frontend/src/app/shared/[shareUuid]/loading.tsx
@@ -1,5 +1,5 @@
 import Section from '~/components/layout/Section'
-import { Skeleton } from '~/components/ui/skeleton/Skeleton'
+import { Skeleton } from '~/components/ui/skeleton'
 
 export default async function Loading() {
   return (
