+++ frontend/src/components/features/restaurant/RestaurantSidebar.tsx
@@ -4,7 +4,7 @@ import type { PageInfo } from '~/types/pagination'
 import type { RestaurantListItem } from '~/types/restaurant'
 import Link from 'next/link'
 import { useEffect, useRef } from 'react'
-import { Button } from '~/ui/buttons/Button'
+import { Button } from '~/components/ui/button'
 import { cn } from '~/utils/cn'
 import RestaurantCard from './RestaurantCard'
 import RestaurantListHeader from './RestaurantListHeader'
@@ -43,48 +43,48 @@ export default function RestaurantSidebar({
       <div className="mt-4 space-y-4">
         {totalCount > 0
           ? (
-              <>
-                {restaurants.map(restaurant => (
-                  <RestaurantCard
-                    key={restaurant.id}
-                    restaurant={restaurant}
+            <>
+              {restaurants.map(restaurant => (
+                <RestaurantCard
+                  key={restaurant.id}
+                  restaurant={restaurant}
+                />
+              ))}
+
+              <div className="mt-4 space-y-2 rounded-lg bg-gray-50/50 p-4">
+                {totalPages > 1 && (
+                  <RestaurantPagination
+                    pagination={pagination}
                   />
-                ))}
+                )}
 
-                <div className="mt-4 space-y-2 rounded-lg bg-gray-50/50 p-4">
-                  {totalPages > 1 && (
-                    <RestaurantPagination
-                      pagination={pagination}
-                    />
-                  )}
-
-                  <p className="text-center text-xs text-muted-foreground">
-                    Powered by
-                    {' '}
-                    <a
-                      href="http://webservice.recruit.co.jp/"
-                      className="text-sky-600"
-                    >
-                      ホットペッパーグルメ Webサービス
-                    </a>
-                  </p>
-                </div>
-              </>
-            )
-          : (
-              <div className="text-center">
-                <p className="text-sm leading-relaxed">
-                  お店が見つかりませんでした。
-                  <br />
-                  条件を変えて再検索してください。
+                <p className="text-center text-xs text-muted-foreground">
+                  Powered by
+                  {' '}
+                  <a
+                    href="http://webservice.recruit.co.jp/"
+                    className="text-sky-600"
+                  >
+                    ホットペッパーグルメ Webサービス
+                  </a>
                 </p>
-                <Link href="/" className="mt-4 inline-block">
-                  <Button>
-                    再検索する
-                  </Button>
-                </Link>
               </div>
-            )}
+            </>
+          )
+          : (
+            <div className="text-center">
+              <p className="text-sm leading-relaxed">
+                お店が見つかりませんでした。
+                <br />
+                条件を変えて再検索してください。
+              </p>
+              <Link href="/" className="mt-4 inline-block">
+                <Button>
+                  再検索する
+                </Button>
+              </Link>
+            </div>
+          )}
       </div>
 
     </div>
