+++ frontend/src/components/ui/autocomplete/StationSuggestions.tsx
@@ -1,6 +1,6 @@
 import type { SavedStation, StationProps } from '~/types/station'
 import Loading from '~/public/figure_loading_circle.svg'
-import { CommandEmpty, CommandGroup, CommandItem, CommandList } from '../command/Command'
+import { CommandEmpty, CommandGroup, CommandItem, CommandList } from '../command'
 
 interface StationSuggestionsProps {
   isLoading: boolean
