+++ frontend/src/components/ui/buttons/AddFormButton.tsx
@@ -1,7 +1,7 @@
 import type { FieldArrayWithId, UseFieldArrayAppend } from 'react-hook-form'
 import type { AreaFormValues } from '~/schemas/station-search.schema'
 import { LuPlus } from 'react-icons/lu'
-import { Button } from './Button'
+import { Button } from '../button'
 
 interface AddFormButtonProps {
   fields: FieldArrayWithId<AreaFormValues>[]
