+++ frontend/src/components/ui/dialogs/ShareFavoriteListDialog.tsx
@@ -6,8 +6,8 @@ import { FaXTwitter } from 'react-icons/fa6'
 import { LuCopy, LuMail, LuShare } from 'react-icons/lu'
 import { toast } from 'sonner'
 import { logger } from '~/lib/logger'
-import { Button } from '../buttons/Button'
-import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from './Dialog'
+import { Button } from '../button'
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from '../dialog'
 
 interface ShareFavoriteListDialogProps {
   isOpen: boolean
