+++ frontend/src/components/ui/forms/LoginForm.tsx
@@ -9,9 +9,9 @@ import { useForm } from 'react-hook-form'
 import { useAuth } from '~/hooks/useAuth'
 import { logger } from '~/lib/logger'
 import { loginSchema } from '~/schemas/login.schema'
-import { Button } from '../buttons/Button'
+import { Button } from '../button'
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from './Form'
-import { Input } from './Input'
+import { Input } from '../input'
 
 interface LoginFormProps {
   onSuccess?: () => void
